requirements/advanced.txt
# Advanced Requirements
# Extended features beyond base functionality
# Install base requirements first: pip install -r base.txt

# ============================================================================
# REDIS - Caching and Rate Limiting
# ============================================================================
redis==5.0.1
aioredis==2.0.1
limits==3.7.0
slowapi==0.1.9

# ============================================================================
# AUTHENTICATION - Two-Factor and Phone Validation
# ============================================================================
pyotp==2.9.0
qrcode==7.4.2
phonenumbers==8.13.27

# ============================================================================
# BACKGROUND TASKS - Celery and Message Queue
# ============================================================================
celery==5.3.4
kombu==5.3.4
flower==2.0.1
pika==1.3.2

# ============================================================================
# EMAIL - Sending and Templating
# ============================================================================
fastapi-mail==1.4.1
jinja2==3.1.3

# ============================================================================
# PAYMENTS - Stripe and PayPal
# ============================================================================
stripe==7.12.0
paypalrestsdk==1.13.1

# ============================================================================
# STORAGE - AWS S3 and Cloud Storage
# ============================================================================
aioboto3==12.3.0
boto3==1.34.10

# ============================================================================
# WEBSOCKET - Real-time Communication
# ============================================================================
python-socketio==5.11.0
aiohttp==3.9.1

# ============================================================================
# MONITORING - Prometheus, Logging, and Tracing
# ============================================================================
prometheus-fastapi-instrumentator==6.1.0
prometheus-client==0.19.0
python-json-logger==2.0.7
structlog==24.1.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0

# ============================================================================
# REPORTING - PDF Generation and Data Processing
# ============================================================================
pandas==2.1.4
numpy==1.26.3
reportlab==4.0.9
weasyprint==60.2
Pillow==10.2.0

# ============================================================================
# UTILITIES - Additional Production Tools
# ============================================================================
python-slugify==8.0.1
pydantic-extra-types==2.4.1
fastapi-versioning==0.10.0
psycopg2-binary==2.9.9
timescale==0.1.1
